# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "AuthenticatedUser": {
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "is_active"
        ],
        "title": "AuthenticatedUser",
        "type": "object"
      },
      "BranchSummary": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "timezone",
          "is_active"
        ],
        "title": "BranchSummary",
        "type": "object"
      },
      "CashSessionView": {
        "properties": {
          "branch_code": {
            "title": "Branch Code",
            "type": "string"
          },
          "branch_id": {
            "format": "uuid",
            "title": "Branch Id",
            "type": "string"
          },
          "branch_name": {
            "title": "Branch Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "opening_amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Opening Amount",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "User Email",
            "type": "string"
          },
          "user_full_name": {
            "title": "User Full Name",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "workstation_code": {
            "title": "Workstation Code",
            "type": "string"
          },
          "workstation_id": {
            "format": "uuid",
            "title": "Workstation Id",
            "type": "string"
          },
          "workstation_name": {
            "title": "Workstation Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "opening_amount",
          "opened_at",
          "branch_id",
          "branch_code",
          "branch_name",
          "workstation_id",
          "workstation_code",
          "workstation_name",
          "user_id",
          "user_email",
          "user_full_name"
        ],
        "title": "CashSessionView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "service",
          "environment",
          "version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        },
        "required": [
          "access_token",
          "user"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "OpenCashSessionRequest": {
        "properties": {
          "opening_amount": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Opening Amount"
          },
          "workstation_code": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Workstation Code",
            "type": "string"
          }
        },
        "required": [
          "workstation_code",
          "opening_amount"
        ],
        "title": "OpenCashSessionRequest",
        "type": "object"
      },
      "PosBootstrapResponse": {
        "properties": {
          "active_cash_session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CashSessionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "branch": {
            "$ref": "#/components/schemas/BranchSummary"
          },
          "local_timestamp": {
            "format": "date-time",
            "title": "Local Timestamp",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          },
          "workstation": {
            "$ref": "#/components/schemas/WorkstationSummary"
          }
        },
        "required": [
          "user",
          "branch",
          "workstation",
          "local_timestamp",
          "active_cash_session"
        ],
        "title": "PosBootstrapResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkstationSummary": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "is_active"
        ],
        "title": "WorkstationSummary",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ZeroMerma API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "me_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/cash-sessions/current": {
      "get": {
        "operationId": "current_cash_session_v1_cash_sessions_current_get",
        "parameters": [
          {
            "in": "query",
            "name": "workstation_code",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Workstation Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CashSessionView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Current Cash Session V1 Cash Sessions Current Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Current Cash Session",
        "tags": [
          "cash-sessions"
        ]
      }
    },
    "/v1/cash-sessions/open": {
      "post": {
        "operationId": "open_cash_session_v1_cash_sessions_open_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenCashSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashSessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Open Cash Session",
        "tags": [
          "cash-sessions"
        ]
      }
    },
    "/v1/pos/bootstrap": {
      "get": {
        "operationId": "bootstrap_v1_pos_bootstrap_get",
        "parameters": [
          {
            "in": "query",
            "name": "workstation_code",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Workstation Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosBootstrapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bootstrap",
        "tags": [
          "pos"
        ]
      }
    }
  }
}